+++ tests/test_client_functional_oldstyle.py
@@ -417,7 +417,7 @@ class TestHttpClientFunctional(unittest.TestCase):
             self.assertEqual(r.status, 200)
 
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
     def test_POST_MULTIPART(self):
         with run_server(self.loop, router=Functional) as httpd:
@@ -447,7 +447,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                 content['multipart-data'][2])
             self.assertEqual(r.status, 200)
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
     def test_POST_STREAM_DATA(self):
         with run_server(self.loop, router=Functional) as httpd:
@@ -475,7 +475,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                     headers={'Content-Length': str(len(data))}))
             content = self.loop.run_until_complete(r.json())
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
             self.assertEqual(str(len(data)),
                              content['headers']['Content-Length'])
@@ -503,7 +503,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                     headers={'Content-Length': str(len(data))}))
             content = self.loop.run_until_complete(r.json())
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
             self.assertEqual(str(len(data)),
                              content['headers']['Content-Length'])
@@ -530,7 +530,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                     headers={'Content-Length': str(len(data))}))
             content = self.loop.run_until_complete(r.json())
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
             self.assertEqual(str(len(data)),
                              content['headers']['Content-Length'])
@@ -559,7 +559,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                     headers={'Content-Length': str(len(data))}))
             content = self.loop.run_until_complete(r.json())
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
             self.assertEqual(str(len(data)),
                              content['headers']['Content-Length'])
@@ -572,7 +572,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                 self.loop.run_until_complete(
                     session.request('get', httpd.url('method', 'get')))
 
-            session.close()
+            self.loop.run_until_complete(session.close())
 
     def test_session_close(self):
         conn = aiohttp.TCPConnector(loop=self.loop)
@@ -594,7 +594,7 @@ class TestHttpClientFunctional(unittest.TestCase):
             self.assertEqual(content['content'], 'requests=1')
             r.close()
 
-        session.close()
+        self.loop.run_until_complete(session.close())
         conn.close()
 
     def test_multidict_headers(self):
@@ -615,7 +615,7 @@ class TestHttpClientFunctional(unittest.TestCase):
             self.assertEqual(str(len(data)),
                              content['headers']['Content-Length'])
 
-        session.close()
+        self.loop.run_until_complete(session.close())
 
     def test_dont_close_explicit_connector(self):
 
@@ -627,7 +627,7 @@ class TestHttpClientFunctional(unittest.TestCase):
             await r.read()
             self.assertEqual(1, len(connector._conns))
             connector.close()
-            session.close()
+            await session.close()
 
         with run_server(self.loop, router=Functional) as httpd:
             url = httpd.url('keepalive')
@@ -669,7 +669,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                 r = await session.request('GET', url)
                 await r.read()
                 self.assertEqual(0, len(connector._conns))
-            session.close()
+            await session.close()
             connector.close()
             server.close()
             await server.wait_closed()
@@ -716,7 +716,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                 await session.request('GET', url)
             self.assertEqual(0, len(connector._conns))
 
-            session.close()
+            await session.close()
             connector.close()
             server.close()
             await server.wait_closed()
@@ -746,7 +746,7 @@ class TestHttpClientFunctional(unittest.TestCase):
             self.assertEqual(
                 content['headers']['Cookie'], 'c1=cookie1; c2=cookie2')
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
     def test_session_headers(self):
         with run_server(self.loop, router=Functional) as httpd:
@@ -764,7 +764,7 @@ class TestHttpClientFunctional(unittest.TestCase):
             self.assertEqual(
                 content['headers']["X-Real-Ip"], "192.168.0.1")
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
     def test_session_headers_merge(self):
         with run_server(self.loop, router=Functional) as httpd:
@@ -787,7 +787,7 @@ class TestHttpClientFunctional(unittest.TestCase):
             self.assertEqual(
                 content['headers']["X-Sent-By"], "aiohttp")
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
     def test_session_auth(self):
         with run_server(self.loop, router=Functional) as httpd:
@@ -803,7 +803,7 @@ class TestHttpClientFunctional(unittest.TestCase):
             self.assertEqual(
                 content['headers']["Authorization"], "Basic bG9naW46cGFzcw==")
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
     def test_session_auth_override(self):
         with run_server(self.loop, router=Functional) as httpd:
@@ -821,7 +821,7 @@ class TestHttpClientFunctional(unittest.TestCase):
                 content['headers']["Authorization"],
                 "Basic b3RoZXJfbG9naW46cGFzcw==")
             r.close()
-            session.close()
+            self.loop.run_until_complete(session.close())
 
     def test_session_auth_header_conflict(self):
         with run_server(self.loop, router=Functional) as httpd:
@@ -833,4 +833,4 @@ class TestHttpClientFunctional(unittest.TestCase):
                 self.loop.run_until_complete(
                     session.request('get', httpd.url('method', 'get'),
                                     headers=headers))
-            session.close()
+            self.loop.run_until_complete(session.close())
